[ai][swift] Remove the instrumentation built on a disproved theory
Three instruments and a frame-drop bound were added on the hypothesis that a
wedged main thread was starving the read loop. The watchdog then disproved it:
the main thread stayed responsive through a 105s stall, and a second gRPC
client later showed the server streaming 14.6MB without stalling at all.

The awaited hand-off returns. It applies natural back-pressure, drops nothing,
and needs no counter; the non-blocking version existed only to stop a stall
that was never happening there. With it go PendingCounter, the depth-3 bound
that discarded frames the renderer would otherwise have drawn, the main-actor
hop timing, the read-loop service timer, and the per-message inter-arrival log.

The watchdog stays, along with the duration formatter it uses, now that it is
the only caller. It is the measurement that settled the question, and the one
worth keeping pointed at the main thread.
